Create game_end method
The game_end method checks if the board is full and asks the player if they would like to play another game in the case of the board being full. It exists the program if the response is not affirmative and creates a new game instance when an affirmative response is received.