Fix game run
The Player class code has been adjusted so that it is not dependent on arrays.

The win_check method now solely checks for a winner.The course of action in the case of a tie is contained in the game_end function.

The player_move method now exists the previous function when recursion takes place.

All other 'Game' class methods but 'gameplay' are now protected.